Add single thread examples
